server/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
Pillow>=10.4.0
python-dotenv>=0.19.0